backend/requirements.txt
fastapi
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
uvicorn
mysql-connector-python
pydantic[email]